Move pkg/ to util/
