+++ config/config.go
@@ -12,7 +12,7 @@ import (
 
 	clientmodel "github.com/prometheus/client_golang/model"
 
-	"github.com/prometheus/prometheus/pkg/strutil"
+	"github.com/prometheus/prometheus/util/strutil"
 )
 
 var (
