+++ promql/test.go
@@ -24,11 +24,11 @@ import (
 
 	clientmodel "github.com/prometheus/client_golang/model"
 
-	"github.com/prometheus/prometheus/pkg/strutil"
-	"github.com/prometheus/prometheus/pkg/testutil"
 	"github.com/prometheus/prometheus/storage"
 	"github.com/prometheus/prometheus/storage/local"
 	"github.com/prometheus/prometheus/storage/metric"
+	"github.com/prometheus/prometheus/util/strutil"
+	"github.com/prometheus/prometheus/util/testutil"
 )
 
 var (
