+++ storage/local/storage_test.go
@@ -24,8 +24,8 @@ import (
 
 	clientmodel "github.com/prometheus/client_golang/model"
 
-	"github.com/prometheus/prometheus/pkg/testutil"
 	"github.com/prometheus/prometheus/storage/metric"
+	"github.com/prometheus/prometheus/util/testutil"
 )
 
 func TestFingerprintsForLabelMatchers(t *testing.T) {
