+++ template/template.go
@@ -27,8 +27,8 @@ import (
 
 	clientmodel "github.com/prometheus/client_golang/model"
 
-	"github.com/prometheus/prometheus/pkg/strutil"
 	"github.com/prometheus/prometheus/promql"
+	"github.com/prometheus/prometheus/util/strutil"
 )
 
 // A version of vector that's easier to use from templates.
