+++ util/stats/timer.go
