+++ web/api/query.go
@@ -26,7 +26,7 @@ import (
 
 	clientmodel "github.com/prometheus/client_golang/model"
 
-	"github.com/prometheus/prometheus/pkg/httputil"
+	"github.com/prometheus/prometheus/util/httputil"
 )
 
 // Enables cross-site script calls.
